Added a test to OptionalWhenBraces rule
