+++ detekt-rules/src/main/kotlin/io/gitlab/arturbosch/detekt/rules/empty/EmptyRule.kt
@@ -3,15 +3,12 @@ package io.gitlab.arturbosch.detekt.rules.empty
 import io.gitlab.arturbosch.detekt.api.CodeSmell
 import io.gitlab.arturbosch.detekt.api.Config
 import io.gitlab.arturbosch.detekt.api.Debt
-import io.gitlab.arturbosch.detekt.api.DetektVisitor
 import io.gitlab.arturbosch.detekt.api.Entity
 import io.gitlab.arturbosch.detekt.api.Issue
 import io.gitlab.arturbosch.detekt.api.Rule
 import io.gitlab.arturbosch.detekt.api.Severity
 import io.gitlab.arturbosch.detekt.rules.asBlockExpression
-import org.jetbrains.kotlin.com.intellij.openapi.util.Key
-import org.jetbrains.kotlin.com.intellij.psi.PsiComment
-import org.jetbrains.kotlin.psi.KtBlockExpression
+import io.gitlab.arturbosch.detekt.rules.hasCommentInside
 import org.jetbrains.kotlin.psi.KtExpression
 
 /**
@@ -29,15 +26,5 @@ abstract class EmptyRule(config: Config) : Rule(config) {
 		blockExpression?.statements?.let {
 			if (it.isEmpty() && !blockExpression.hasCommentInside()) report(CodeSmell(issue, Entity.from(this)))
 		}
-	}
-
-	private fun KtBlockExpression.hasCommentInside(): Boolean {
-		val commentKey = Key<Boolean>("comment")
-		this.acceptChildren(object : DetektVisitor() {
-			override fun visitComment(comment: PsiComment?) {
-				if (comment != null) putUserData(commentKey, true)
-			}
-		})
-		return getUserData(commentKey) ?: false
 	}
 }
