+++ detekt-rules/src/test/kotlin/io/gitlab/arturbosch/detekt/rules/style/OptionalWhenBracesSpec.kt
@@ -1,7 +1,7 @@
 package io.gitlab.arturbosch.detekt.rules.style
 
 import io.gitlab.arturbosch.detekt.test.lint
-import org.assertj.core.api.Assertions
+import org.assertj.core.api.Assertions.assertThat
 import org.jetbrains.spek.api.dsl.describe
 import org.jetbrains.spek.api.dsl.it
 import org.jetbrains.spek.subject.SubjectSpek
@@ -11,19 +11,33 @@ class OptionalWhenBracesSpec : SubjectSpek<OptionalWhenBraces>({
 
 	describe("check optional braces in when expression") {
 
+		it("has necessary braces") {
+			val code = """
+				fun x() {
+					when (1) {
+						1 -> foo
+						2 -> {
+							foo()
+							bar()
+						}
+						else -> {
+							// a comment
+							foo()
+						}
+					}
+				}"""
+			assertThat(subject.lint(code)).hasSize(0)
+		}
+
 		it("has unnecessary braces") {
 			val code = """
 				fun x() {
 					when (1) {
 						1 -> { foo() }
-						2 -> foo
-						3 -> {
-							foo()
-							bar()
-						}
+						else -> bar()
 					}
 				}"""
-			Assertions.assertThat(subject.lint(code)).hasSize(1)
+			assertThat(subject.lint(code)).hasSize(1)
 		}
 	}
 })
